Revert "changed prices to BigDecimal / new field in Reservation / test fixes wip / statistics fix / new column / added temp validator for reservation price"
This reverts commit cf6d006385d3d5f2bdfcb4180c55ec12a5e23bd5.

+++ src/main/java/CPSF/com/demo/controller/CamperPlaceController.java
@@ -7,7 +7,6 @@ import CPSF.com.demo.request.CamperPlaceRequest;
 import CPSF.com.demo.service.CamperPlaceService;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -23,7 +22,7 @@ public class CamperPlaceController {
 
     @PostMapping("/create")
     public void createCamperPlace(@RequestBody CamperPlaceRequest request) {
-        camperPlaceService.create(request.type(), BigDecimal.valueOf(request.price()));
+        camperPlaceService.create(request.type(), request.price());
     }
 
     @GetMapping("/getAll")
