+++ src/test/java/CPSF/com/demo/reservationCalculatorTest/ReservationCalculatorTest.java
@@ -3,8 +3,6 @@ package CPSF.com.demo.reservationCalculatorTest;
 import CPSF.com.demo.util.ReservationCalculator;
 import org.junit.jupiter.api.Test;
 
-import java.math.BigDecimal;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
@@ -14,18 +12,18 @@ public class ReservationCalculatorTest {
 
 	@Test
 	public void shouldReturnCorrectCostWithNoDiscount() {
-		BigDecimal camperPlacePrice = BigDecimal.valueOf(100);
-		BigDecimal expected = BigDecimal.valueOf(300);
-		BigDecimal actual = reservationCalculator.calculateFinalReservationCost(3, camperPlacePrice);
+		double camperPlacePrice = 100;
+		double expected = 300;
+		double actual = reservationCalculator.calculateFinalReservationCost(3, camperPlacePrice);
 
 		assertEquals(expected, actual);
 	}
 
 	@Test
 	public void shouldReturnCorrectCostWithDiscount() {
-		BigDecimal camperPlacePrice = BigDecimal.valueOf(100);
-		BigDecimal expected = BigDecimal.valueOf(930);
-		BigDecimal actual = reservationCalculator.calculateFinalReservationCost(10, camperPlacePrice);
+		double camperPlacePrice = 100;
+		double expected = 930;
+		double actual = reservationCalculator.calculateFinalReservationCost(10, camperPlacePrice);
 
 		assertEquals(expected, actual);
 	}
